requirements.txt
# Core Django
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Authentication & Security
djangorestframework-simplejwt==5.3.1
django-allauth==0.61.1
dj-rest-auth==5.0.2
PyJWT==2.8.0

# Password Security
argon2-cffi==23.1.0

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Environment Variables
python-decouple==3.8
python-dotenv==1.0.0

# Google Cloud Services
google-cloud-storage==2.14.0
google-cloud-secret-manager==2.17.0

# Permissions & Security
django-guardian==2.4.0
django-ratelimit==4.1.0
django-cryptography==1.1

# Audit Logging
django-auditlog==2.3.0

# Notifications
fcm-django==2.0.0
firebase-admin==6.4.0

# Email
sendgrid==6.11.0

# Future Payments (Install when ready)
# stripe==8.2.0
# dj-stripe==2.8.3

# API Documentation
drf-spectacular==0.27.1

# Testing
pytest==7.4.4
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==22.6.0

# Code Quality
black==24.1.1
flake8==7.0.0
isort==5.13.2

# Production Server
gunicorn==21.2.0
whitenoise==6.6.0

# Monitoring & Error Tracking
sentry-sdk==1.40.0

# Utilities
pillow==10.2.0
python-dateutil==2.8.2
pytz==2024.1
celery==5.3.6
redis==5.0.1
openai~=2.12.0